+++ tests/clients/test_slack_client.py
@@ -35,13 +35,33 @@ def user_dict():
             "real_name_normalized": "John Doe",
             "display_name_normalized": "Johnny",
             "team": "my-team",
+            "email": "john@my-team.org",
+        },
+    }
+
+
+@pytest.fixture
+def user_dict_no_email():
+    return {
+        "id": "U1",
+        "name": "john",
+        "is_bot": False,
+        "updated": 0,
+        "is_app_user": False,
+        "profile": {
+            "avatar_hash": "abc",
+            "real_name": "John Doe",
+            "display_name": "Johnny",
+            "real_name_normalized": "John Doe",
+            "display_name_normalized": "Johnny",
+            "team": "my-team",
         },
     }
 
 
 @pytest.fixture
 def user(user_dict):
-    return User.parse_obj(user_dict)
+    return User.model_validate(user_dict)
 
 
 @pytest.fixture
@@ -51,7 +71,7 @@ def channel_dict():
 
 @pytest.fixture
 def channel(channel_dict):
-    return Channel.parse_obj(channel_dict)
+    return Channel.model_validate(channel_dict)
 
 
 @pytest.fixture
@@ -121,7 +141,23 @@ async def test_process_users_channels_team_join(slack_client, socket_mode_client
     slack_client._on_team_join.assert_called_once_with(event)
     assert len(slack_client._users) == 1
     assert "U1" in slack_client._users
-    assert slack_client._users["U1"] == User.parse_obj(user_dict)
+    assert slack_client._users["U1"] == User.model_validate(user_dict)
+    assert len(slack_client._users_by_email) == 1
+    assert "john@my-team.org" in slack_client._users_by_email
+    assert slack_client._users_by_email["john@my-team.org"] == User.model_validate(user_dict)
+
+
+@pytest.mark.asyncio
+async def test_process_users_channels_team_join_no_email(slack_client, socket_mode_client, user_dict_no_email):
+    event = {"type": "team_join", "user": user_dict_no_email}
+    req = create_req(event)
+    await slack_client._process_users_channels(socket_mode_client, req)
+    socket_mode_client.send_socket_mode_response.assert_called_once()
+    slack_client._on_team_join.assert_called_once_with(event)
+    assert len(slack_client._users) == 1
+    assert "U1" in slack_client._users
+    assert slack_client._users["U1"] == User.model_validate(user_dict_no_email)
+    assert len(slack_client._users_by_email) == 0
 
 
 @pytest.mark.asyncio
@@ -133,7 +169,7 @@ async def test_process_users_channels_user_change(slack_client, socket_mode_clie
     slack_client._on_user_change.assert_called_once_with(event)
     assert len(slack_client._users) == 1
     assert "U1" in slack_client._users
-    assert slack_client._users["U1"] == User.parse_obj(user_dict)
+    assert slack_client._users["U1"] == User.model_validate(user_dict)
 
 
 @pytest.mark.asyncio
@@ -146,7 +182,7 @@ async def test_process_users_channels_channel_created(slack_client, socket_mode_
     slack_client._on_channel_created.assert_called_once_with(event)
     assert len(slack_client._channels) == 1
     assert "C1" in slack_client._channels
-    assert slack_client._channels["C1"] == Channel.parse_obj(channel_dict)
+    assert slack_client._channels["C1"] == Channel.model_validate(channel_dict)
 
 
 @pytest.mark.asyncio
@@ -201,3 +237,19 @@ async def test_process_users_channels_channel_archive(
     assert len(slack_client._channels) == 1
     assert "C1" in slack_client._channels
     assert slack_client._channels["C1"].is_archived is True
+
+
+@pytest.mark.asyncio
+async def test_get_user_by_id(slack_client, socket_mode_client, user_dict):
+    event = {"type": "team_join", "user": user_dict}
+    req = create_req(event)
+    await slack_client._process_users_channels(socket_mode_client, req)
+    assert slack_client.get_user_by_id("U1") == User.model_validate(user_dict)
+
+
+@pytest.mark.asyncio
+async def test_get_user_by_email(slack_client, socket_mode_client, user_dict):
+    event_with_email = {"type": "team_join", "user": user_dict}
+    req_with_email = create_req(event_with_email)
+    await slack_client._process_users_channels(socket_mode_client, req_with_email)
+    assert slack_client.get_user_by_email("john@my-team.org") == User.model_validate(user_dict)